+++ rollup.config.js
@@ -1,5 +1,5 @@
 import babel from 'rollup-plugin-babel';
-import uglify from 'rollup-plugin-uglify';
+import { uglify } from 'rollup-plugin-uglify';
 import { minify } from 'uglify-es';
 
 const { NODE_ENV } = process.env;
@@ -10,11 +10,13 @@ const external = Object.keys(globals);
 const babelOptions = {
   babelrc: false,
   presets: [
-    ['env', { modules: false }],
-    'stage-3',
+    ['@babel/preset-env', { modules: false }],
   ],
   plugins: [
-    'external-helpers',
+    '@babel/plugin-syntax-dynamic-import',
+    '@babel/plugin-syntax-import-meta',
+    ['@babel/plugin-proposal-class-properties', { loose: false }],
+    '@babel/plugin-proposal-json-strings'
   ],
   exclude: 'node_modules/**',
 };