+++ src/propagation/gstime.js
@@ -33,10 +33,9 @@ import { jday } from '../ext';
 function gstimeInternal(jdut1) {
   const tut1 = (jdut1 - 2451545.0) / 36525.0;
 
-  let temp =
-    (-6.2e-6 * tut1 * tut1 * tut1) +
-    (0.093104 * tut1 * tut1) +
-    (((876600.0 * 3600) + 8640184.812866) * tut1) + 67310.54841; // # sec
+  let temp = (-6.2e-6 * tut1 * tut1 * tut1)
+    + (0.093104 * tut1 * tut1)
+    + (((876600.0 * 3600) + 8640184.812866) * tut1) + 67310.54841; // # sec
   temp = ((temp * deg2rad) / 240.0) % twoPi; // 360/86400 = 1/240, to deg, to rad
 
   //  ------------------------ check quadrants ---------------------